refactor: Move di of mappers to separate file and remove date parser.
