+++ app/src/main/java/com/amsterdam/aflami/di/appModule.kt
@@ -3,6 +3,7 @@ package com.amsterdam.aflami.di
 val appModule = listOf(
     localDataSourceModule,
     remoteDataSourceModule,
+    mapperModule,
     repositoryModule,
     useCaseModule,
     viewModelModule,
