--- repository/src/main/java/com/example/repository/mapper/DateParser.kt
@@ -1,5 +1,0 @@
-package com.example.repository.mapper
-
-interface DateParser {
-    fun parseYear(date: String): Int
-}
