+++ repository/src/main/java/com/example/repository/utils/DateParser.kt
@@ -1,0 +1,7 @@
+package com.example.repository.utils
+
+class DateParser {
+    fun parseYear(date: String): Int {
+        return date.takeIf { it.length >= 4 }?.substring(0, 4)?.toIntOrNull() ?: 0
+    }
+}
